move all 'map features' related pages to their own subfolder
